Move annotation.tsx to annotator folder
